+++ packages/components/src/Label/Label.test.tsx
@@ -2,6 +2,7 @@ import { render, screen } from '@testing-library/react';
 
 import { Theme, ThemeProvider, cva } from '@marigold/system';
 
+import { FieldGroup } from '../FieldBase';
 import { Label } from './Label';
 
 const theme: Theme = {
@@ -55,22 +56,12 @@ test('renders <label> element by default', () => {
   expect(label instanceof HTMLLabelElement).toBeTruthy();
 });
 
-test('can render as <span>', () => {
-  render(
-    <ThemeProvider theme={theme}>
-      <Label as="span">label</Label>
-    </ThemeProvider>
-  );
-  const label = screen.getByText(/label/);
-  expect(label instanceof HTMLSpanElement).toBeTruthy();
-});
-
 test('accepts labelwidth as css variable and set the style', () => {
   render(
     <ThemeProvider theme={theme}>
-      <Label as="span" labelWidth="100px">
-        label
-      </Label>
+      <FieldGroup labelWidth="100px">
+        <Label>label</Label>
+      </FieldGroup>
     </ThemeProvider>
   );
   const label = screen.getByText(/label/);